Rename test files Move some common test code to module
